FIX: High CPU usage fixes (#6)
* FIX: If sendAllArtDMX is run for a node, it results in high CPU usage

When the instance is not a Controller (and hence does not use sendAllArtDMX), sendAllArtDMX runs in a tight loop without sleeps or waits. There is no need for this task to run at all for non-controllers.

* FIX: High CPU usage waiting for SendArtPollEvent

Tight loop caused high CPU usage, instead simply delay until the poll period has elapsed

* FIX: sendAllArtDMX is also called by nodes

---------

Co-authored-by: Patrick <patrick-dmxc@users.noreply.github.com>